Change push constant member to uint32_t

+++ include/nbl/builtin/hlsl/sampling/hierarchical_image/common.hlsl
@@ -18,7 +18,7 @@ struct SLumaGenPushConstants
 {
   uint32_t lumaMapWidth : 16;
   uint32_t lumaMapHeight : 16;
-  uint16_t lumaMapLayer;
+  uint32_t lumaMapLayer;
 };
 
 struct SWarpGenPushConstants
@@ -28,7 +28,7 @@ struct SWarpGenPushConstants
   uint32_t warpMapWidth : 16;
   uint32_t warpMapHeight : 16;
   // Both warpMap and lumaMap should have the same layer count
-  uint16_t lumaMapLayer;
+  uint32_t lumaMapLayer;
 };
 
 NBL_CONSTEXPR_INLINE_NSPC_SCOPE_VAR uint32_t GenWarpWorkgroupDim = 16;
